Crate hyperquark

Modules

error

instructions

Contains information about instructions (roughly anaologus to blocks), including input type validation, output type mapping, and WASM generation.

ir

optimisation

prelude

commonly used things which would be nice not to have to type out every time

rc

registry

sb3

1-1 representation of project.json or sprite.json files in the sb3 format.

wasm

Macros

hq_assert

hq_assert_eq

hq_bad_proj

hq_bug

hq_todo

index_counter

instructions_test

Generates unit tests for instructions files.

log

make_hq_bad_proj

for use in ok_or_else and similar methods

make_hq_bug

for use in ok_or_else and similar methods

make_hq_todo

for use in ok_or_else and similar methods

memory_layout

warn

Structs

HQError

Enums

HQErrorType

Functions

log

sb3_to_wasm

warn

Type Aliases

HQResult