hyperquark::instructions

Module fields

Structs

ControlIfElseFields

ControlLoopFields

ControlWaitFields

DataAddtolistFields

we need these fields to be mutable for optimisations to be feasible

DataDeletealloflistFields

we need these fields to be mutable for optimisations to be feasible

DataDeleteoflistFields

we need these fields to be mutable for optimisations to be feasible

DataInsertatlistFields

we need these fields to be mutable for optimisations to be feasible

DataItemoflistFields

we need these fields to be mutable for optimisations to be feasible

DataLengthoflistFields

we need these fields to be mutable for optimisations to be feasible

DataListcontentsFields

we need these fields to be mutable for optimisations to be feasible

DataReplaceitemoflistFields

we need these fields to be mutable for optimisations to be feasible

DataSetvariabletoFields

we need these fields to be mutable for optimisations to be feasible

DataTeevariableFields

DataVariableFields

DataVisvariableFields

EventBroadcastAndWaitFields

EventBroadcastFields

HqBooleanFields

HqBoxFields

HqCastFields

HqColorRgbFields

HqFloatFields

HqIntegerFields

HqTextFields

HqYieldFields

LooksSayFields

LooksThinkFields

ProceduresArgumentFields

ProceduresCallNonwarpFields

ProceduresCallWarpFields

SensingAskandwaitFields