hyperquark::instructions

Module wrap_instructions

Module for generating WASM for a sequence of instructions.

Some definitions for formality:

Define {$#`is_root` : #`Fork` -> #`bool`$} to be {$ #`is_root`;(f) = f.#`branch`.#`is_none`;(). $}

Let {$F$} be the function mapping a Fork to all Forks reachable from it, with {$F(f)$} a shorthand for {$F(#`Some`;(f))$}:

  F(#`None`)    &= {}\
  F(#`Some`;(f)) &= { f } union
    F(#`Some`;(f.#`branch`?.#`b_if`)) union
    F(f.#`branch`?.#`b_else`).

Let {$L$} be the function mapping a Fork to its leaves:

L(f) = { g in F(f) | #`is_root`;(g) }.

These definitions are used in pre-/post-conditions and invariants for functions and datatypes.

Structs

Branch 🔒

An if/else branch. Invariant {$I_b$} declared here.

ForestBuilderState 🔒

State during build_forest. Invariants {$I_f$} and {$I_f'$} declared here.

Fork 🔒

A block of instructions, possibly ending in a branch, along with information about the type stack at the beginning of the block, and just before the branch

Leaves 🔒

A collection of forks that are leaves. Invariant {$I_l$} declared here.

Roots 🔒

Base, top-level branches across the whole function; i.e. branches that begin with an empty type stack. Invariants {$I_r$} and {$I_r'$} declared here.

Functions

box_type_check 🔒

Returns the list of instructions used to check if a boxed value is of the specified (base) type

boxed_pattern 🔒

Returns the NaN-box bit pattern for the provided type (which must be basic)

build_boxed_locals 🔒

build_forest 🔒

Builds a ‘forest’ (collection of trees, in this case ordered) representing the WASM for the specified opcodes, with the entry point having the specified type stack.

collapse_same_typed_leaves 🔒

If all of the leaves have the same exit stack, we can pre-emptively merge them together to avoid a huge blow up in the number of leaves.

make_branch 🔒

Makes a new branch for type selection, using the latest_fork as the else branch, and a new leaf as the if branch. Adds the new leaf to new_leaves.

make_branch_fork 🔒

Precondition: new_ty.is_base_type() Postcondition: latest_fork.is_some()

make_leaf 🔒

make_type_branches 🔒

Postcondition: new_leaves.is_some()

request_screen_refresh 🔒

should_unbox 🔒

Should arguments to this opcode be unboxed?

unbox_instructions 🔒

Returns the list of instructions used to unbox a boxed value of the provided (base) type

wrap_instructions

Type Aliases

BoxedLocals 🔒