hyperquark::instructions::wrap_instructions

Struct ForestBuilderState

struct ForestBuilderState {
    this_root: Rc<Fork>,
    roots: Roots,
    leaves: Leaves,
}

State during build_forest. Invariants {$I_f$} and {$I_f'$} declared here.

Invariant {$I_f$}

#`roots`;[#`roots`.#`len`;() - 1] = #`this_root`.

Invariant {$I_f'$}

#`leaves` = L(#`this_root`).

Fields
this_root: Rc<Fork>roots: Rootsleaves: Leaves
Implementations

impl ForestBuilderState

fn add_new_root_if_needed(&mut self)

Adds a new root if the exit type stack on the latest tree is empty.

Maintains invariant {$I_r$}: a new root can only be added if the previous root has all leaves with empty type stacks. Inductively {$I_r$} must hold for any value of roots.

impl ForestBuilderState

fn new(type_stack: Rc<[IrType]>) -> Self