hyperquark::ir::step

Struct Step

pub struct Step {
    context: StepContext,
    opcodes: Vec<IrOpcode>,
    used_non_inline: bool,
    id: Box<str>,
    project: Weak<IrProject>,
}

Fields
context: StepContextopcodes: Vec<IrOpcode>used_non_inline: boolid: Box<str>

used for Hash. Should be obtained from a block in the Step where possible.

project: Weak<IrProject>
Implementations

impl Step

pub const fn context(&self) -> &StepContext

pub const fn opcodes(&self) -> &Vec<IrOpcode>

pub const fn used_non_inline(&self) -> bool

pub const fn id(&self) -> &str

pub fn project(&self) -> Weak<IrProject>

pub fn new( id: Option<Box<str>>, context: StepContext, opcodes: Vec<IrOpcode>, project: Weak<IrProject>, used_non_inline: bool, ) -> Self

pub fn new_empty( project: Weak<IrProject>, used_non_inline: bool, target: Rc<Target>, ) -> Self

Creates a totally empty noop step. This should not be used outside of wasm module generation.

pub fn new_terminating( context: StepContext, project: Weak<IrProject>, used_non_inline: bool, ) -> Self

pub fn new_poll_waiting_threads( context: StepContext, project: Weak<IrProject>, ) -> Self

pub fn new_poll_waiting_event( context: StepContext, project: Weak<IrProject>, ) -> Self

pub fn new_poll_timer(context: StepContext, project: Weak<IrProject>) -> Self

pub const fn opcodes_mut(&mut self) -> &mut Vec<IrOpcode>

pub fn from_block( block: &Block, block_id: Box<str>, blocks: &BlockMap, context: &StepContext, project: &Weak<IrProject>, final_next_blocks: NextBlocks, used_non_inline: bool, flags: &WasmFlags, ) -> HQResult<Self>

pub fn from_block_non_inlined( block: &Block, block_id: Box<str>, blocks: &BlockMap, context: &StepContext, project: &Weak<IrProject>, final_next_blocks: NextBlocks, flags: &WasmFlags, ) -> HQResult<StepIndex>

pub fn make_inlined(&mut self)

pub fn clone_to_non_inlined( &self, project: &Weak<IrProject>, ) -> HQResult<StepIndex>

pub fn does_yield(&self) -> bool

pub fn globally_scoped_variables( &self, ) -> HQResult<impl DoubleEndedIterator<Item = RcVar> + Clone>

An iterator of variables that are global (in the WASM sense, not the scratch sense) in this step. Should be in a consistent order. Used for procedures.

pub fn globally_scoped_variables_num(&self) -> HQResult<usize>