hyperquark::ir

Struct IrType

#[repr(transparent)]
pub struct IrType {
    bits: u32,
}

a bitmask of possible IR types

Fields
bits: u32
Implementations

impl Type

pub const IntZero: Type

pub const IntPos: Type

pub const IntNeg: Type

pub const IntNonZero: Type

pub const Int: Type

pub const FloatPosZero: Type

pub const FloatNegZero: Type

pub const FloatZero: Type

pub const FloatPosInt: Type

pub const FloatPosFrac: Type

pub const FloatPosReal: Type

pub const FloatNegInt: Type

pub const FloatNegFrac: Type

pub const FloatNegReal: Type

pub const FloatPosInf: Type

pub const FloatNegInf: Type

pub const FloatInf: Type

pub const FloatNan: Type

pub const FloatPos: Type

pub const FloatNeg: Type

pub const FloatPosWhole: Type

pub const FloatNegWhole: Type

pub const FloatInt: Type

pub const FloatFrac: Type

pub const FloatReal: Type

pub const FloatNotNan: Type

pub const Float: Type

pub const BooleanTrue: Type

pub const BooleanFalse: Type

pub const Boolean: Type

pub const QuasiInt: Type

pub const Number: Type

pub const StringNumber: Type

pub const StringBoolean: Type

pub const StringNan: Type

pub const String: Type

pub const QuasiBoolean: Type

pub const QuasiNumber: Type

pub const Any: Type

pub const ColorRGB: Type

pub const ColorARGB: Type

pub const Color: Type

pub const AnyOrColor: Type

pub fn flags() -> impl Iterator<Item = &'static (&'static str, Self)>

Returns an iterator over all flags of the bitmask. Where each Item = (name, flag).

pub const fn bits(&self) -> u32

Returns the underlying bits of the bitmask.

pub const fn all_bits() -> Self

Returns a bitmask that contains all values.

This will include bits that do not have any flags. Use ::all_flags() if you only want to use flags.

pub const fn all_flags() -> Self

Returns a bitmask that contains all flags.

pub const fn is_all_bits(&self) -> bool

Returns true if the bitmask contains all values.

This will check for bits == !0, use .is_all_flags() if you only want to check for all flags

pub const fn is_all_flags(&self) -> bool

Returns true if the bitmask contains all flags.

This will fail if any unused bit is set, consider using .truncate() first.

pub const fn all() -> Self

👎Deprecated:
Please use the ::all_bits() constructor

Returns a bitmask that contains all values.

This will include bits that do not have any flags. Use ::all_flags() if you only want to use flags.

pub const fn is_all(&self) -> bool

👎Deprecated:
Please use the .is_all_bits() method

Returns true if the bitmask contains all values.

This will check for bits == !0, use .is_all_flags() if you only want to check for all flags

pub const fn full() -> Self

👎Deprecated:
Please use the ::all_flags() constructor

Returns a bitmask that contains all flags.

pub const fn is_full(&self) -> bool

👎Deprecated:
Please use the .is_all_flags() method

Returns true if the bitmask contains all flags.

This will fail if any unused bit is set, consider using .truncate() first.

pub const fn none() -> Self

Returns a bitmask that does not contain any values.

pub const fn is_none(&self) -> bool

Returns true if the bitmask does not contain any values.

pub const fn truncate(&self) -> Self

Returns a bitmask that only has bits corresponding to flags

pub const fn intersects(&self, other: Self) -> bool

Returns true if self intersects with any value in other, or if other does not contain any values.

This is equivalent to (self & other) != 0 || other == 0.

pub const fn contains(&self, other: Self) -> bool

Returns true if self contains all values of other.

This is equivalent to (self & other) == other.

pub const fn not(self) -> Self

Returns the bitwise NOT of the bitmask.

pub const fn and(self, other: Self) -> Self

Returns the bitwise AND of the bitmask.

pub const fn or(self, other: Self) -> Self

Returns the bitwise OR of the bitmask.

pub const fn xor(self, other: Self) -> Self

Returns the bitwise XOR of the bitmask.

impl Type

pub const BASE_TYPES: [Self; 6]

pub fn is_base_type(self) -> bool

pub fn base_type(self) -> Option<Self>

pub fn base_types(self) -> Box<dyn Iterator<Item = Self>>

pub const fn maybe_positive(self) -> bool

pub const fn maybe_negative(self) -> bool

pub const fn maybe_zero(self) -> bool

pub const fn maybe_nan(self) -> bool

pub const fn maybe_inf(self) -> bool

pub const fn none_if_false(condition: bool, if_true: Self) -> Self