hyperquark::ir

Struct StepContext

pub struct StepContext {
    pub target: Rc<Target>,
    pub warp: bool,
    pub proc_context: Option<ProcContext>,
    pub debug: bool,
}

Fields
target: Rc<Target>warp: bool

whether or not the current thread is warped. this may be because the current procedure is warped, or because a procedure higher up the call stack was warped.

proc_context: Option<ProcContext>debug: bool

enables certain behaviours such as console.log say/think rather than displaying in bubbles

Implementations

impl StepContext

pub fn project(&self) -> HQResult<Rc<IrProject>>

pub const fn target(&self) -> &Rc<Target>