hyperquark::optimisation

Module ssa

Carry out SSA (static single assignment) transformation on the project, and fix up some variable shennanigans.

The goals are:

Split variables so they are only ever written to once, and are read from the most recent instance of the variable

Change variable reads inside procedures to argument access, if that variable hasn’t been written to inside that procedure yet

Assign the narrowest type possible to each variable

Reinsert casts to account for newly assigned variable types.

Box procedure returns if needed

This will create a lot of unnecessary variables; this is ok, as wasm-opt does local merging and should remove unnecessary tees etc. But this might be an area in which performance could be improved, if we can reduce the number of variables we need to track on our end, which will also improve wasm-opt’s runtime.

Rough description of algorithm

This does not correspond exactly to the implementation, but hopefully explains the intent.

function split_to_SSA (step, ssa_map = {}, mut additional_opcodes):
  for each opcode in step:
    if opcode is a variable read and is not already local:
      if opcode not in ssa_map:
        if step context is in procedure:
          replace opcode with the appropriate argument access
      else:
        replace opcode with local variable read to corresponding variable in ssa_map
    if opcode is a variable write/tee to var and is not already local:
      new_var := new variable
      ssa_map[var] = new_var
      replace opcode with local variable write/tee to new_var
    if opcode yields inline to other_steps (e.g. by yield_inline, loop or if/else):
      if opcode is loop:
        for each globally accessible variable:
          write to a new SSA with the current value of the variable
          # the needless SSAs here will be removed by wasm-opt.
          # TODO: can we keep track of which variables are accessed from inside each step, to simplify this step?
      lower_ssa_maps := []
      for each lower_step in other_steps:
        append (split_to_SSA(other step, ssa_map, additional_opcodes)) to lower_ssa_maps
      consolidate lower_ssa_maps into lower_ssa_map of variable : [variable]
      for each (global, locals) in lower_ssa_map:
        if locals.length == 1:
          ssa_map[global] := locals[0]
        else:
          new_var := new variable
          for each i, lower_step in enumerate(other_steps):
            append (local variable write to new_var from locals[i])
                to additional_opcodes[lower_step]
          ssa_map[global] = new_var
    if opcode yields non-inline to other_step:
      split_to_SSA(other_step, {}, additional_steps)
  if (step is not inlined and step context is not in procedure) or (last opcode was a non-inline yield):
    for (global, local) in ssa_map:
      append (local variable read of local) to additional_opcodes[step]
      append (global variable write of global) to additional_opcodes[step]
    return {}
  if step context is in procedure:
    return {}
  return ssa_map

additional_opcodes = {}

for step in top level steps:
  split_to_SSA (step, {}, additional_opcodes)

for (step, extra_opcodes) in additional_opcodes:
  concat extra_opcodes to step.opcodes

Alongside this SSA splitting, we also need to construct a graph to reason about what types might be on the stack at any one point. This is simpler so won’t be explained here.

A note on naming: I call this SSA, but it is not strictly true. In SSA, each variable is assigned to exactly once, and at a dominance frontier, SSAs from preceding blocks are merged into one using a ‘phi node’ or ‘phi function’, which picks the correct variable to read based on which branch was just taken. Carrying around additional information about which block was just taken isn’t something that I want to do, so our ‘phi function’ actually inserts the relevant assignment into the preceding blocks directly. This does mean that some variables can be written to from different blocks, especially where loops are concerned (as the output SSA needs to be written to both before the loop and at the end of each iteration). This isn’t actually a problem for us, as we’re not using SSA to carry out liveness analysis - we use it to minimise casts and maximise the type information known at compile time.

Modules

box_proc_returns 🔒

type_convergence 🔒

var_graph 🔒

Structs

SSAToken

A token type that cannot be instantiated from anywhere else (since the field is private)

Functions

optimise_variables

split_variables_and_make_graphs 🔒

Splits variables at every write site, and ensures that outer/global variables are then written to at the end of a step. This also constructs variable type graphs for each step which can then be analyzed to determine the best types for variables.

visit_procedure 🔒

visit_step_recursively 🔒