hyperquark

Module prelude

commonly used things which would be nice not to have to type out every time

Re-exports

pub use crate::rc::Rc;

pub use crate::rc::Weak;

pub use crate::registry::NamedRegistrar;

pub use crate::registry::NamedRegistry;

pub use crate::registry::NamedRegistryItem;

pub use crate::registry::NamedRegistryItemOverride;

pub use crate::registry::Registry;

pub use crate::registry::RegistryDefault;

pub use crate::registry::RegistryType;

pub use crate::registry::TryNamedRegistryItemOverride;

pub use crate::HQError;

pub use crate::HQResult;

Modules

fmt

Utilities for formatting and printing strings.

Structs

BTreeMap

An ordered map based on a B-Tree.

BTreeSet

An ordered set based on a B-Tree.

Box

A pointer type that uniquely owns a heap allocation of type T.

PhantomPinned

A marker type which does not implement Unpin.

Pin

A pointer which pins its pointee in place.

RefCell

A mutable memory location with dynamically checked borrow rules

String

A UTF-8–encoded, growable string.

Vec

A contiguous growable array type, written as Vec<T>, short for ‘vector’.

Enums

Cow

A clone-on-write smart pointer.

Traits

Borrow

A trait for borrowing data.

Itertools

An Iterator blanket implementation that provides extra adaptors and methods.

ToString

A trait for converting a value to a String.

Type Aliases

IndexMap

IndexSet