hyperquark::registry

Struct NamedRegistry

pub struct NamedRegistry<R>(MapRegistry<R::Key, R::Value>)
where
    R: NamedRegistrar<R>;

Registry type for items that will be registered with an arbitrary “name” of some sort which is constable.

This means we can store the value statically in a trait-implementing struct, and derive the “name” from that struct (by default from the name of the struct).

A NameRegistry is defined with respect to a NamedRegistrar, which can be any object (probably a ZST) that implements RegistryType, defining the types Key (defaults to Box<str>) and Value. Manually implementing NamedRegistrar is only necessary if Key: !From<&str> or you wish to specify a custom naming function.

To use a ZST T as an itemin a NamedRegistry, implement NamedRegistryItem<Value> for T to define the default value for the item.

If you wish items in this registry to be overridable, impl NamedRegistryItemOverride<Value, A> for T, where A is an argument to pass to the r#override function. If multiple arguments are needed, they should be wrapped into a tuple or struct.

Examples

struct NumberRegistrar;
impl RegistryType for NumberRegistrar {
    type Key = Box<str>;
    type Value = u32;
}
// `NamedRegistrar` is implemented automatically because `Key = Box<str>`

type NumberRegistry = NamedRegistry<NumberRegistrar>;

struct Zero;
impl NamedRegistryItem<u32> for Zero {
    const VALUE: u32 = 0;
}

struct Natural;
impl NamedRegistryItem<u32> for Natural {
    const VALUE: u32 = 1;
}
impl NamedRegistryItemOverride<u32, u32> for Natural {
    fn r#override(new: u32) -> u32 {
        new
    }
}

let num_reg = NumberRegistry::default();
let nat_pos: usize = num_reg.register::<Natural, _>()?;
let zero_pos: usize = num_reg.register::<Zero, _>()?;
let also_nat_pos: usize = num_reg.register_override::<Natural, _, _>(5)?;
assert_eq!(nat_pos, also_nat_pos);

// let also_zero_pos: usize = num_reg.register_override::<Zero, _, _>(0)?;
// ^ would fail to compile because `Zero` doesn't implement `NamedRegistryItemOverride`

Tuple Fields
0: MapRegistry<R::Key, R::Value>
Implementations

impl NamedRegistry<StaticFunctionRegistrar>

pub fn finish( self, functions: &mut FunctionSection, exports: &mut ExportSection, codes: &mut CodeSection, type_registry: &TypeRegistry, imported_func_count: u32, ) -> HQResult<()>

impl NamedRegistry<TableRegistrar>

pub fn finish(self, tables: &mut TableSection, exports: &mut ExportSection)

impl<R> NamedRegistry<R>
where R: NamedRegistrar<R>,

pub fn registry(&self) -> &RefCell<IndexMap<R::Key, R::Value>>

pub fn register<T, N>(&self) -> HQResult<N>
where N: TryFrom<usize>, <N as TryFrom<usize>>::Error: Debug, T: NamedRegistryItem<R::Value>,

Registers a NamedRegistryItem using its key function and its const VALUE

pub fn register_dyn<N>(&self, key: R::Key, value: R::Value) -> HQResult<N>
where N: TryFrom<usize>, <N as TryFrom<usize>>::Error: Debug,

Registers a runtime key-value pair; just calls register on the underlying Registry

pub fn register_dyn_override<N>( &self, key: R::Key, value: R::Value, ) -> HQResult<N>
where N: TryFrom<usize>, <N as TryFrom<usize>>::Error: Debug,

Override a runtime key-value pair with a runtime value; just calls register_override on the underlying Registry

pub fn register_override<T, N, A>(&self, override_arg: A) -> HQResult<N>
where N: TryFrom<usize>, <N as TryFrom<usize>>::Error: Debug, T: NamedRegistryItem<R::Value> + NamedRegistryItemOverride<R::Value, A>,

Overrides a NamedRegistryItem entry using the override argument types associated with the corresponding NamedRegistryItemOverride

pub fn try_register_override<T, N, A>(&self, override_arg: A) -> HQResult<N>
where N: TryFrom<usize>, <N as TryFrom<usize>>::Error: Debug, T: NamedRegistryItem<R::Value> + TryNamedRegistryItemOverride<R::Value, A>,

Tries to override a NamedRegistryItem entry using the override argument types associated with the corresponding TryNamedRegistryItemOverride