hyperquark::sb3

Enum BlockOpcode

pub enum BlockOpcode {

    control_repeat,
    control_repeat_until,
    control_while,
    control_for_each,
    control_forever,
    control_wait,
    control_wait_until,
    control_if,
    control_if_else,
    control_stop,
    control_create_clone_of,
    control_delete_this_clone,
    control_get_counter,
    control_incr_counter,
    control_clear_counter,
    control_all_at_once,
    control_start_as_clone,
    control_create_clone_of_menu,
    data_variable,
    data_setvariableto,
    data_changevariableby,
    data_hidevariable,
    data_showvariable,
    data_listcontents,
    data_addtolist,
    data_deleteoflist,
    data_deletealloflist,
    data_insertatlist,
    data_replaceitemoflist,
    data_itemoflist,
    data_itemnumoflist,
    data_lengthoflist,
    data_listcontainsitem,
    data_hidelist,
    data_showlist,
    event_broadcast,
    event_broadcast_menu,
    event_broadcastandwait,
    event_whenflagclicked,
    event_whenkeypressed,
    event_whenthisspriteclicked,
    event_whentouchingobject,
    event_whenstageclicked,
    event_whenbackdropswitchesto,
    event_whengreaterthan,
    event_whenbroadcastreceived,
    looks_say,
    looks_sayforsecs,
    looks_think,
    looks_thinkforsecs,
    looks_show,
    looks_hide,
    looks_hideallsprites,
    looks_switchcostumeto,
    looks_switchbackdropto,
    looks_switchbackdroptoandwait,
    looks_nextcostume,
    looks_nextbackdrop,
    looks_changeeffectby,
    looks_seteffectto,
    looks_cleargraphiceffects,
    looks_changesizeby,
    looks_setsizeto,
    looks_changestretchby,
    looks_setstretchto,
    looks_gotofrontback,
    looks_goforwardbackwardlayers,
    looks_size,
    looks_costumenumbername,
    looks_backdropnumbername,
    looks_costume,
    looks_backdrops,
    motion_movesteps,
    motion_gotoxy,
    motion_goto,
    motion_goto_menu,
    motion_turnright,
    motion_turnleft,
    motion_pointindirection,
    motion_pointtowards,
    motion_glidesecstoxy,
    motion_glideto,
    motion_ifonedgebounce,
    motion_setrotationstyle,
    motion_changexby,
    motion_setx,
    motion_changeyby,
    motion_sety,
    motion_xposition,
    motion_yposition,
    motion_direction,
    motion_scroll_right,
    motion_scroll_up,
    motion_align_scene,
    motion_xscroll,
    motion_yscroll,
    motion_pointtowards_menu,
    operator_add,
    operator_subtract,
    operator_multiply,
    operator_divide,
    operator_lt,
    operator_equals,
    operator_gt,
    operator_and,
    operator_or,
    operator_not,
    operator_random,
    operator_join,
    operator_letter_of,
    operator_length,
    operator_contains,
    operator_mod,
    operator_round,
    operator_mathop,
    pen_clear,
    pen_stamp,
    pen_penDown,
    pen_penUp,
    pen_setPenColorToColor,
    pen_changePenColorParamBy,
    pen_setPenColorParamTo,
    pen_changePenSizeBy,
    pen_setPenSizeTo,
    pen_setPenShadeToNumber,
    pen_changePenShadeBy,
    pen_setPenHueToNumber,
    pen_changePenHueBy,
    pen_menu_colorParam,
    procedures_definition,
    procedures_call,
    procedures_prototype,
    argument_reporter_string_number,
    argument_reporter_boolean,
    sensing_touchingobject,
    sensing_touchingcolor,
    sensing_coloristouchingcolor,
    sensing_distanceto,
    sensing_distancetomenu,
    sensing_timer,
    sensing_resettimer,
    sensing_of,
    sensing_mousex,
    sensing_mousey,
    sensing_setdragmode,
    sensing_mousedown,
    sensing_keypressed,
    sensing_current,
    sensing_dayssince2000,
    sensing_loudness,
    sensing_loud,
    sensing_askandwait,
    sensing_answer,
    sensing_username,
    sensing_userid,
    sensing_touchingobjectmenu,
    sensing_keyoptions,
    sensing_of_object_menu,
    sound_play,
    sound_playuntildone,
    sound_stopallsounds,
    sound_seteffectto,
    sound_changeeffectby,
    sound_cleareffects,
    sound_sounds_menu,
    sound_beats_menu,
    sound_effects_menu,
    sound_setvolumeto,
    sound_changevolumeby,
    sound_volume,

}

A possible block opcode.

Encompasses the default block pallette, the pen extension, and a few hidden but non-obsolete blocks. A block being listed here does not imply that it is supported by HyperQuark.

Variants

control_repeat

control_repeat_until

control_while

control_for_each

control_forever

control_wait

control_wait_until

control_if

control_if_else

control_stop

control_create_clone_of

control_delete_this_clone

control_get_counter

control_incr_counter

control_clear_counter

control_all_at_once

control_start_as_clone

control_create_clone_of_menu

data_variable

data_setvariableto

data_changevariableby

data_hidevariable

data_showvariable

data_listcontents

data_addtolist

data_deleteoflist

data_deletealloflist

data_insertatlist

data_replaceitemoflist

data_itemoflist

data_itemnumoflist

data_lengthoflist

data_listcontainsitem

data_hidelist

data_showlist

event_broadcast

event_broadcast_menu

event_broadcastandwait

event_whenflagclicked

event_whenkeypressed

event_whenthisspriteclicked

event_whentouchingobject

event_whenstageclicked

event_whenbackdropswitchesto

event_whengreaterthan

event_whenbroadcastreceived

looks_say

looks_sayforsecs

looks_think

looks_thinkforsecs

looks_show

looks_hide

looks_hideallsprites

looks_switchcostumeto

looks_switchbackdropto

looks_switchbackdroptoandwait

looks_nextcostume

looks_nextbackdrop

looks_changeeffectby

looks_seteffectto

looks_cleargraphiceffects

looks_changesizeby

looks_setsizeto

looks_changestretchby

looks_setstretchto

looks_gotofrontback

looks_goforwardbackwardlayers

looks_size

looks_costumenumbername

looks_backdropnumbername

looks_costume

looks_backdrops

motion_movesteps

motion_gotoxy

motion_goto

motion_goto_menu

motion_turnright

motion_turnleft

motion_pointindirection

motion_pointtowards

motion_glidesecstoxy

motion_glideto

motion_ifonedgebounce

motion_setrotationstyle

motion_changexby

motion_setx

motion_changeyby

motion_sety

motion_xposition

motion_yposition

motion_direction

motion_scroll_right

motion_scroll_up

motion_align_scene

motion_xscroll

motion_yscroll

motion_pointtowards_menu

operator_add

operator_subtract

operator_multiply

operator_divide

operator_lt

operator_equals

operator_gt

operator_and

operator_or

operator_not

operator_random

operator_join

operator_letter_of

operator_length

operator_contains

operator_mod

operator_round

operator_mathop

pen_clear

pen_stamp

pen_penDown

pen_penUp

pen_setPenColorToColor

pen_changePenColorParamBy

pen_setPenColorParamTo

pen_changePenSizeBy

pen_setPenSizeTo

pen_setPenShadeToNumber

pen_changePenShadeBy

pen_setPenHueToNumber

pen_changePenHueBy

pen_menu_colorParam

procedures_definition

procedures_call

procedures_prototype

argument_reporter_string_number

argument_reporter_boolean

sensing_touchingobject

sensing_touchingcolor

sensing_coloristouchingcolor

sensing_distanceto

sensing_distancetomenu

sensing_timer

sensing_resettimer

sensing_of

sensing_mousex

sensing_mousey

sensing_setdragmode

sensing_mousedown

sensing_keypressed

sensing_current

sensing_dayssince2000

sensing_loudness

sensing_loud

sensing_askandwait

sensing_answer

sensing_username

sensing_userid

sensing_touchingobjectmenu

sensing_keyoptions

sensing_of_object_menu

sound_play

sound_playuntildone

sound_stopallsounds

sound_seteffectto

sound_changeeffectby

sound_cleareffects

sound_sounds_menu

sound_beats_menu

sound_effects_menu

sound_setvolumeto

sound_changevolumeby

sound_volume