hyperquark::wasm::func

Struct StepFunc

pub struct StepFunc {
    locals: RefCell<Vec<ValType>>,
    available_locals: RefCell<BTreeMap<ValType, BTreeSet<u32>>>,
    instructions: RefCell<Vec<Instruction>>,
    params: Box<[ValType]>,
    output: Box<[ValType]>,
    registries: Rc<Registries>,
    flags: WasmFlags,
    local_variables: RefCell<BTreeMap<RcVar, u32>>,
    target: StepTarget,
    target_index: u32,
    costume_names: Rc<Vec<Vec<Box<str>>>>,
}

representation of a step’s function

Fields
locals: RefCell<Vec<ValType>>available_locals: RefCell<BTreeMap<ValType, BTreeSet<u32>>>instructions: RefCell<Vec<Instruction>>params: Box<[ValType]>output: Box<[ValType]>registries: Rc<Registries>flags: WasmFlagslocal_variables: RefCell<BTreeMap<RcVar, u32>>target: StepTargettarget_index: u32costume_names: Rc<Vec<Vec<Box<str>>>>
Implementations

impl StepFunc

pub fn registries(&self) -> Rc<Registries>

pub const fn flags(&self) -> &WasmFlags

pub const fn instructions(&self) -> &RefCell<Vec<Instruction>>

pub fn params(&self) -> &[ValType]

pub fn output(&self) -> &[ValType]

pub const fn target(&self) -> StepTarget

pub const fn target_index(&self) -> u32

pub const fn costume_names(&self) -> &Rc<Vec<Vec<Box<str>>>>

pub fn new( registries: Rc<Registries>, flags: WasmFlags, target: StepTarget, target_index: u32, costume_names: Rc<Vec<Vec<Box<str>>>>, ) -> Self

creates a new step function, with one paramter

pub fn new_with_types( params: Box<[ValType]>, output: Box<[ValType]>, registries: Rc<Registries>, flags: WasmFlags, target: StepTarget, target_index: u32, costume_names: Rc<Vec<Vec<Box<str>>>>, ) -> Self

creates a new step function with the specified amount of paramters. currently only used in testing to validate types

pub fn local_variable(&self, var: &RcVar) -> HQResult<u32>

pub fn local(&self, val_type: ValType) -> HQResult<u32>

Registers a new local in this function, and returns its index

pub fn free_local(&self, index: u32) -> HQResult<()>

pub fn add_instructions( &self, instructions: impl IntoIterator<Item = Instruction>, ) -> HQResult<()>

pub fn finish( self, funcs: &mut FunctionSection, code: &mut CodeSection, events: &BTreeMap<Event, Vec<u32>>, types: &Rc<TypeRegistry>, threads_count_global: u32, spawn_new_thread_func: u32, spawn_thread_in_stack_func: u32, threads_table: u32, imported_func_count: u32, static_func_count: u32, imported_global_count: u32, ) -> HQResult<()>

Takes ownership of the function and returns the backing wasm_encoder Function

fn compile_instructions( &self, opcodes: &[IrOpcode], ) -> HQResult<Vec<Instruction>>

pub fn compile_step( step: &RefCell<Step>, step_index: StepIndex, steps: &Rc<RefCell<Vec<Self>>>, registries: Rc<Registries>, flags: WasmFlags, costume_names: Rc<Vec<Vec<Box<str>>>>, ) -> HQResult<Self>

pub fn compile_inner_step<S>(&self, step: S) -> HQResult<Vec<Instruction>>
where S: Deref<Target = RefCell<Step>>,