hyperquark::wasm::mem_layout

Module sprite

Constants

BLOCK_SIZE

COSTUME

current costume number, 0-indexed (i32)

PEN_BRIGHTNESS

value component of pen colour (0-100) (f32)

PEN_COLOR

hue component of pen colour (0-100) (f32)

PEN_COLOR_A

alpha component of rgba representation of pen colour (0-1) (f32)

PEN_COLOR_B

blue component of rgba representation of pen colour (0-1) (f32)

PEN_COLOR_G

green component of rgba representation of pen colour (0-1) (f32)

PEN_COLOR_R

red component of rgba representation of pen colour (0-1) (f32)

PEN_DOWN

non-zero if pen down, 0 otherwise (i8)

PEN_SATURATION

saturation component of pen colour (0-100) (f32)

PEN_SIZE

radius of pen (f64)

PEN_TRANSPARENCY

transparency of pen (f32)

ROTATION

sprite rotation, in scratch angles (0 = up, 90 = right) (f64)

SIZE

sprite size, where default is 100(%) (f64)

VISIBLE

non-zero if sprite is visible, 0 otherwise (i8)

X

x-position of sprite (f64)

Y

y-position of sprite (f64)

_PADDING

bytes 58-59 padding